requirements.txt
pytorch-lightning>=1.5.6
tensorboard>=2.2.0
torchsummary>=1.5.1